Function b2dp::mechanisms::exponential::exponential_mechanism

pub fn exponential_mechanism<T, R: ThreadRandGen + Copy, F: Fn(&T) -> f64>(
    eta: Eta, 
    outcomes: &Vec<T>, 
    utility: F, 
    utility_min: i64, 
    utility_max: i64, 
    max_outcomes: u32, 
    rng: R, 
    options: ExponentialOptions
) -> Result<&T, &'static str>

Implements the base-2 exponential mechanism. Utility convention is to take -utility(o), and utility_min is therefore the highest possible weight/maximum probability outcome. This mechanism does not scale based on the sensitivity of the utility function. For a utility function with sensitivity alpha, the mechanism is 2*alpha*eta base-2 DP, and 2*alpha*ln(2)*eta base-e DP.
The caller must ensure that utility_min, utility_max, max_outcomes and outcomes are determined independently of the utility function and any private data.

Arguments

eta: the base-2 privacy parameter

outcomes: the set of outcomes the mechanism chooses from

utility: utility function operating on elements of outcomes. utility does not explicitly take a database input, and is expected to have a pointer to the database or access to the private data needed to determine utilities.

utility_min: the minimum utility permitted by the mechanism (highest possible weight)

utility_max: the maximum utility permitted by the mechanism (lowest possible weight)

max_outcomes: the maximum number of outcomes permitted by the mechanism

rng: a random number generator

Returns

Returns a reference to an element in outcomes sampled according to the base-2 exponential mechanism.

Known Timing Channels

This mechanism has known timing channels. Please see normalized_sample.

Errors

Returns Err if any of the parameters are configured incorrectly or if inexact arithmetic occurs.

Example

use b2dp::{exponential_mechanism, Eta, GeneratorOpenSSL, ExponentialOptions};
 
fn util_fn (x: &u32) -> f64 {
    return ((*x as f64)-0.0).abs();
}
let eta = Eta::new(1,1,1).unwrap(); 
let utility_min = 0;
let utility_max = 10;
let max_outcomes = 10;
let rng = GeneratorOpenSSL {};
let options = ExponentialOptions {min_retries: 1, optimized_sample: true, empirical_precision: false};
let outcomes: Vec<u32> = (0..max_outcomes).collect();
let result = exponential_mechanism(eta, &outcomes, util_fn, 
                                    utility_min, utility_max, 
                                    max_outcomes,
                                    rng, options);

Exact Arithmetic

This function calls enter_exact_scope() and exit_exact_scope(), and therefore clears the mpfr::flags and does not preserve the incoming flag state.