Function b2dp::mechanisms::integerpartition::integer_partition_mechanism_with_bounds

pub fn integer_partition_mechanism_with_bounds(
    eta: Eta, 
    x: &Vec<i64>, 
    pb: &PartitionBound, 
    options: IntegerPartitionOptions
) -> Result<Vec<i64>>

The integer partition mechanism given bounds

Arguments

eta: privacy parameter

x: the target (private) integer partition

pb: the partition bounds

options: mechanism options

Returns

Returns a vector representing a private approximation of the target integer partition or an error if the partition could not be approximated.

Privacy Budget

Consumes 2*eta in privacy budget for single add/remove adjacency.

Exact Arithmetic

This function calls enter_exact_scope() and exit_exact_scope(), and therefore clears the mpfr::flags and does not preserve the incoming flag state. This method may need to be re-tried (or integer_partition_mechanism_with_weights may be used) with higher precision on failure.

Timing Channels

This method increases precision as needed dependign on private data, and should be treated as unsafe for timing channels.