Function b2dp::mechanisms::integerpartition::integer_partition_mechanism_with_weights

pub fn integer_partition_mechanism_with_weights(
    weight_table: &mut WeightTable, 
    pb: &PartitionBound, 
    options: IntegerPartitionOptions
) -> Result<Vec<i64>>

The integer partition mechanism given a pre-computed weight table.

Arguments

weight_table: a pre-computed weight table for the integer partition exponential mechanism.

partition_bounds:

Returns

Returns a vector representing a private approximation of the target integer partition or an error if the partition could not be approximated.

Privacy Budget

(Should be accounted for in weight computation.)

Exact Arithmetic

This function calls enter_exact_scope() and exit_exact_scope(), and therefore clears the mpfr::flags and does not preserve the incoming flag state. This method may need to be re-tried with higher precision on failure.

Timing Channels

This mechanism has known timing channels. Please see normalized_sample.