Function b2dp::mechanisms::laplace::clamped_laplace_mechanism

pub fn clamped_laplace_mechanism<R: ThreadRandGen + Copy>(
    eta: Eta, 
    lower_bound: f64, 
    upper_bound: f64, 
    target: f64, 
    gamma: f64, 
    rng: R, 
    options: ExponentialOptions
) -> Result<f64>

Implements clamped Laplace mechanism via the base-2 exponential mechanism Outputs an f64 in the range [lower_bound, upper_bound] of the form lower_bound + gamma*i.

Arguments

eta: the base-2 privacy parameter

lower_bound: the lower bound of the range for the Laplace mechanism

upper_bound: the upper bound of the range for the Laplace mechanism

target: the target value (value to approximate)

gamma: the granularity of the outcome space

rng: a random number generator

Returns

Returns an f64 within the specified bounds.

Errors

Returns Err if any of the parameters are configured incorrectly or if inexact arithmetic occurs.

Example

Sample from the Laplace mechanism targeting 0.0 clamped to the range [-10,10] with granularity 0.25.

use b2dp::{mechanisms::laplace::clamped_laplace_mechanism, Eta, GeneratorOpenSSL};
 
let target = 0.0;
let eta = Eta::new(1,1,1).unwrap(); 
let lower_bound = -10.0;
let upper_bound = 10.0;
let gamma = 0.25;
let rng = GeneratorOpenSSL {};
let sample = clamped_laplace_mechanism(eta, lower_bound, upper_bound, target, gamma, rng, Default::default()).unwrap();