Function b2dp::mechanisms::sparsevector::sparse_vector

pub fn sparse_vector<R: ThreadRandGen + Copy>(
    eta1: Eta, 
    eta2: Eta, 
    c: usize, 
    query_values: &Vec<f64>, 
    gamma: f64, 
    query_min: f64, 
    query_max: f64, 
    w: f64, 
    rng: R, 
    optimize: bool
) -> Result<Vec<bool>>

The sparse vector mechanism Sensitivity (Delta) is assumed to be 1. Privacy parameters must be scaled appropriately if this is not the case.

Arguments:

eta1: the privacy budget to spend on sampling the threshold

eta2: the privacy budget to spend on the queries. Sensitivity of queries assumed to be 1. eta2 must be scaled appropriately for c.

c: the number of allowed positive queries.

gamma: the granularity (must be a reciprocal of a positive integer)

Q: a set of query values which must be integer multiples of gamma, otherwise they are rounded, which must be accounted for in privacy budget calculations.

Qmin: the minimum value any query may be.

Qmax: the maximum value any query may be.

w: the width to use for threshold comparison (higher width increases accuracy at the cost of efficiency, but only up to a point)

rng: randomness source

optimize: whether to optimize sampling (exacerbates timing channels)

Privacy budget usage

Uses eta1*Delta + 2*eta2*Delta*c. As such, eta and eta2 should be scaled appropriately for the desired usage.

Returns

A vector of bool with at most c true entries indicating which queries exceeded the noisy threshold or an error if insufficient precision to do so.

Exact Arithmetic

This method calls enter_exact_scope which clears the mpfr::flags.

Timing Channels

This method uses normalized_sample which has known timing channels. Care should be taken given that the method allows for an arbitrary number of queries which may amplify timing channels.