Module b2dp::utilities::discretesampling

Implements discrete sampling methods in base-2 including lazy_threshold and sample_within_bounds. Function names/signatures are not finalized and are likely to be changed.

Functions

adjust_eta
Adjust the given eta to account for granularity 1/gamma_inv.

conditional_lazy_threshold
Determines whether the discretized Laplace exceeds the given threshold conditioned on it exceeding the given conditional threshold.

get_sum
Returns the sum: (1-base)*\sum_{k=start}^{end}base^{|k|}

is_multiple_of
Returns whether a is an integer multiple of b.

lazy_threshold
Determines whether the discretized Laplace exceeds the given threshold.

sample_within_bounds
Samples from the Discrete Laplace mechanism at granularity gamma within the provided bounds wmin and wmax where the values wmin and wmax are sampled with probability equal to the sum of the probabilities of all elements greater than or less than the bounds, respectively.