Function b2dp::utilities::discretesampling::adjust_eta

pub fn adjust_eta(
    eta: Eta, 
    gamma_inv: &Float, 
    arithmeticconfig: &mut ArithmeticConfig
) -> Result<Eta, &'static str>

Adjust the given eta to account for granularity 1/gamma_inv

Arguments

eta: privacy parameter

gamma_inv: an integer-valued Float representing the inverse of gamma

arithmeticconfig: an ArithmeticConfig with sufficient precision to adjust eta if it can be adjusted.

Returns

A new Eta adjusted such that 2^eta_prime = (2^eta)^gamma or an error if eta cannot be adjusted.

Exact Arithmetic

Does not enforce exact arithmetic, this is the caller's responsibility.