Function b2dp::utilities::discretesampling::get_sum

pub fn get_sum(
    base: &Float, 
    arithmeticconfig: &ArithmeticConfig, 
    start: Float, 
    end: Float
) -> Result<Float, &'static str>

Returns the sum: 0.5*(1-base)^{-1}*\sum_{k=start}^{end}base^{|k|}

Arguments:

base: a Float indicating the base for the sum

arithmeticconfig: an ArithmeticConfig

start: an integer-valued or infinite Float indicating the starting point of the sum

end: an integer-valued or infinite Float indicating the ending point of the sum

Returns

Either the sum or an error if parameters are mis-specified. This method does not explicitly check for inexact arithmetic, it is the caller's responsibility to do so.

Exact Arithmetic

This method does not enforce exact arithmetic, this is the caller's responsibility.

Timing channels

The recursive calls to get_sum introduce a timing channel distinguishing whether start and end cross zero.