Module b2dp::utilities::exactarithmetic

Implements methods requiring exact arithmetic, and encapsulates all unsafe code to access mpfr::flags to determine whether computations are exact.

Structs

ArithmeticConfig
The exact arithmetic configuration. Includes the precision of all mechanism arithmetic and status bits indicating if any inexact arithmetic has been performed. The ArithmeticConfig implementation encapsulates all unsafe calls to mpfr.

Functions

normalized_sample
Normalized Weighted Sampling Returns the index of the element sampled according to the weights provided. Uses optimized sampling if optimize set to true. Setting optimize to true exacerbates timing channels.

randomized_round
Randomized Rounding