Function b2dp::utilities::exactarithmetic::normalized_sample

pub fn normalized_sample<R: ThreadRandGen>(
    weights: &Vec<Float>, 
    arithmetic_config: &mut ArithmeticConfig, 
    rng: R, 
    optimize: bool
) -> Result<usize, &'static str>

Normalized Weighted Sampling Returns the index of the element sampled according to the weights provided. Uses optimized sampling if optimize set to true. Setting optimize to true exacerbates timing channels.

Arguments

weights: the set of weights to use for sampling; all weights must be positive, zero-weight elements are not permitted.

arithmetic_config: the arithmetic config specifying precision

rng: source of randomness.

optimize: whether to optimize sampling, introducing a timing channel and an error condition side channel.

Returns

Returns an index of an element sampled according to the weights provided. If the precision of the provided ArithmeticConfig is insufficient for sampling, the method returns an error. Note that errors are not returned on inexact arithmetic, and the caller is responsible for calling enter_exact_scope() and exit_exact_scope() to monitor inexact arithmetic.

Known Timing Channels

This method has known timing channels. They result from: (1) Generating a random value in [0,2^k] and (2) (In optimized sampling only) To determine the index corresponding to the random value, the method iterates through cumulative weights and terminates the loop when the index is found and (3) (In optimized sampling only) Checking for zero weights These can be exploited in several ways:

Rejection probability: if the adversary can control the total weight of the utilities such that the probability of rejection in the random index generation stage changes, the time needed for sampling will vary between adjacent databases. The difference in time will depend on the speed of random number generation. By default, ArithmeticConfig sets the minimum retries to 1. To reduce the probability that this timing channel is accessible to an adversary, the minimum number of retries can be increased via ArithmeticConfig::set_retries.

Optimized sampling:

Ordering of weights: if the adversary can change the ordering of the weights such that the largest weights (most probable) weights are first under a certain condition, and the largest weights are last if that condition doesn't hold, then the adversary can use the time of normalized_sample to guess whether the condition holds.

Size of weights: if the adversary can change the size of the weights such that if a certain condition holds, the weight is more concentrated and if not the weight is less concentrated, then the adversary can use the time taken by normalized_sample as a signal for whether the condition holds.

Zero weight: optimized sampling also rejects immediately if a zero weight is encountered. If the adversary can inject a zero weight at a particular position in the weights depending on a private condition, they can use the time it takes to return an error as a timing channel.

The timing channels for optimized sampling could be somewhat (but not completely) mitigated by shuffling the weights prior to calling normalized_sample.

Exact Arithmetic

normalized_sample does not explicitly call enter_exact_scope() or exit_exact_scope(), and therefore preserves any mpfr::flags that are set before the function is called.