Function b2dp::utilities::exactarithmetic::randomized_round

pub fn randomized_round<R: ThreadRandGen>(
    x: f64, 
    arithmetic_config: &mut ArithmeticConfig, 
    rng: R
) -> i64

Randomized Rounding

Arguments

x: the value to round

arithmetic_config: the arithmetic configuration to use

Returns

x rounded to the nearest smaller or larger integer by drawing a random value rho in [0,1] and rounding down if rho > x_fract, rounding up otherwise.