Struct b2dp::utilities::weights::WeightTable

pub struct WeightTable {
    pub weights: HashMap<Key, Float>,
    pub arithmetic_config: ArithmeticConfig,
    pub coefficients: HashMap<Key, Float>,
    pub total_weights: HashMap<Key, Float>,
    pub probabilities: HashMap<Key, Float>,
}

Weight table for sampling from the integer partition exponential mechanism and computing.

Fields
weights: HashMap<Key, Float>

The weight table

arithmetic_config: ArithmeticConfig

Arithmetic config

coefficients: HashMap<Key, Float>

Coefficients table for bias computation

total_weights: HashMap<Key, Float>

Total weights table for bias computation

probabilities: HashMap<Key, Float>

Probabilities table for bias computation

Methods

impl WeightTable

pub fn from_bounds(
eta: Eta,
pb: &PartitionBound,
x: &Vec<i64>
) -> Result<WeightTable, &'static str>

Computes the weight table for the given bounds, enforcing an exact bound if exact total specified in exact.

Arguments

eta: privacy parameter

pb: a set of upper and lower bounds for the partition

x: the target (private) integer partition

Returns

A weight table consisting of the cumulative weights for all completions beginning with value q at index i.

Exact Arithmetic

This function calls enter_exact_scope() and exit_exact_scope(), and therefore clears the mpfr::flags and does not preserve the incoming flag state.

Known Timing Channels

This method does not compute worst-case arithmetic up front, and instead increases precision as needed during computation.

pub fn inexact_from_bounds(
eta: Eta,
pb: &PartitionBound,
x: &Vec<i64>
) -> Result<WeightTable, &'static str>

DO NOT USE except for bug demonstration! Computes the weight table for the given bounds WITHOUT exact arithmetic enforcement.

Arguments

eta: privacy parameter

pb: a set of upper and lower bounds for the partition

x: the target (private) integer partition

Returns

DO NOT USE. A weight table consisting of the cumulative weights for all completions beginning with value q at index i, with INTENTIONALLY INEXACT ARITHMETIC for comparison and testing purposes only.

pub fn get_bias(
&mut self,
pb: &PartitionBound,
x: &Vec<i64>
) -> Result<Vec<f64>, &'static str>

Computes the bias of sampling from the weight table. Stores helper tables of coefficients, total weights, etc, for later use if needed.

Arguments

pb: the PartitionBound to be used for sampling

x: the target partition x. (Note that this does not have to be the same target partition that was used to create the table.)

Returns

A vector indicating the bias of each index if x were the target partition.